tests.pkg: fix handling of architecture specific packages
The tests.pkg file, when using an apt backend, would easily get confused
when a test had installed a package from another architecture, such as
i386, as the architecture would be discarded by the logic. This could
easily lead to attempt to remove the amd64 version of the package,
causing utter havoc.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>